+++ vertx3-jdk8/src/main/kotlin/uy/klutter/vertx/Vertx.kt
@@ -6,27 +6,23 @@ import nl.komponents.kovenant.Deferred
 import nl.komponents.kovenant.Kovenant
 import nl.komponents.kovenant.Promise
 import nl.komponents.kovenant.deferred
-import java.time.LocalDateTime
-import java.time.OffsetDateTime
-import java.time.ZonedDateTime
-import java.time.temporal.Temporal
 import kotlin.reflect.KClass
-import kotlin.reflect.jvm.java
 
 
-/**
+/*
  *  VertxInit.ensure() is called from methods that startup vertx to be sure Kovenant and Vertx are configured to work
  *  together.  If you do not startup vertx using these methods, you should directly call VertxInit.ensure() at the start
- *  of your application.  Not all vertx+Kovenant methods call VertxInit.ensure, although those that will not affect performance
- *  might.
+ *  of your application.
  */
 
-public class WrappedThrowableException(cause: Throwable): Exception(cause.message, cause)
+class WrappedThrowableException(cause: Throwable) : Exception(cause.message, cause)
 
 /**
- * Start vert.x returning a Kovenant Promise<Vertx, Exception>
+ * Start vert.x
+ *
+ * @return Promise<Vertx, Exception>
  */
-public fun vertx(): Promise<Vertx, Exception> {
+fun vertx(): Promise<Vertx, Exception> {
     VertxInit.ensure()
 
     val deferred = deferred<Vertx, Exception>()
@@ -41,9 +37,19 @@ public fun vertx(): Promise<Vertx, Exception> {
 }
 
 /**
- * Start vert.x returning a Kovenant Promise<Vertx, Exception>
+ * Start vert.x (alias for [`vertx()`][vertx] function)
+ *
+ * @return Promise<Vertx, Exception>
  */
-public fun vertx(options: VertxOptions): Promise<Vertx, Exception> {
+fun promiseVertx(): Promise<Vertx, Exception> = vertx()
+
+/**
+ * Start vert.x
+ *
+ * @param options configuration options for vert.x
+ * @return Promise<Vertx, Exception>
+ */
+fun vertx(options: VertxOptions): Promise<Vertx, Exception> {
     VertxInit.ensure()
 
     val deferred = deferred<Vertx, Exception>()
@@ -51,217 +57,232 @@ public fun vertx(options: VertxOptions): Promise<Vertx, Exception> {
         deferred.resolve(Vertx.vertx(options))
     } catch (ex: Exception) {
         deferred.reject(ex)
-    }  catch (ex: Throwable) {
+    } catch (ex: Throwable) {
         deferred.reject(WrappedThrowableException(ex))
     }
     return deferred.promise
 }
 
 /**
- * Start clustered vert.x returning a Kovenant Promise<Vertx, Exception>
+ * Start vert.x (alias for [`vertx(options)`][vertx] function)
+ *
+ * @param options configuration options for vert.x
+ * @return Promise<Vertx, Exception>
  */
-public fun vertxCluster(options: VertxOptions): Promise<Vertx, Exception> {
-    VertxInit.ensure()
+fun promiseVertx(options: VertxOptions): Promise<Vertx, Exception> = vertx(options)
 
-    val deferred = deferred<Vertx, Exception>()
-    Vertx.clusteredVertx(options, promiseResult(deferred))
-    return deferred.promise
+/**
+ * Start clustered vert.x
+ *
+ * @param options configuration options for vert.x
+ * @return Promise<Vertx, Exception>
+ */
+fun vertxCluster(options: VertxOptions): Promise<Vertx, Exception> {
+    return withDeferred { Vertx.clusteredVertx(options, promiseResult(it)) }
 }
+
+/**
+ * Start clustered vert.x (alias for [`vertxCluster(options)`][vertxCluster] function)
+ *
+ * @param options configuration options for vert.x
+ * @return Promise<Vertx, Exception>
+ */
+fun promiseVertxCluster(options: VertxOptions): Promise<Vertx, Exception> = vertxCluster(options)
 
 /**
  * retrieve the current vert.x context if one is attached to the current thread
  */
-public fun vertxContext(): Context? = Vertx.currentContext()
+fun vertxContext(): Context? = Vertx.currentContext()
 
 /**
- * Deploy a verticle, returning a Promise<deploymentId as String, Exception>
+ * Deploy a verticle
+ *
+ * @param verticle to deploy
+ * @return Promise<deploymentId as String, Exception>
  */
-public fun Vertx.promiseDeployVerticle(verticle: Verticle): Promise<String, Exception> {
-    VertxInit.ensure()
-
-    val deferred = deferred<String, Exception>()
-    deployVerticle(verticle, promiseResult(deferred))
-    return deferred.promise
+fun Vertx.promiseDeployVerticle(verticle: Verticle): Promise<String, Exception> {
+    return withDeferred { deployVerticle(verticle, promiseResult(it)) }
 }
 
 /**
- * Deploy a verticle, returning a Promise<deploymentId as String, Exception>
+ * Deploy a verticle
+ *
+ * @param verticle to deploy
+ * @param options deployment options
+ * @return Promise<deploymentId as String, Exception>
  */
-public fun Vertx.promiseDeployVerticle(verticle: Verticle, options: DeploymentOptions): Promise<String, Exception> {
-    VertxInit.ensure()
-
-    val deferred = deferred<String, Exception>()
-    deployVerticle(verticle, options, promiseResult(deferred))
-    return deferred.promise
+fun Vertx.promiseDeployVerticle(verticle: Verticle, options: DeploymentOptions): Promise<String, Exception> {
+    return withDeferred { deployVerticle(verticle, options, promiseResult(it)) }
 }
 
 /**
- * Deploy a verticle, returning a Promise<deploymentId as String, Exception>
+ * Deploy a verticle
+ *
+ * @param verticleClass reference to Verticle class (Kotlin `KClass`) to deploy
+ * @return Promise<deploymentId as String, Exception>
  */
-public fun  <T : AbstractVerticle> Vertx.promiseDeployVerticle(verticleClass: KClass<T>): Promise<String, Exception> {
-    VertxInit.ensure()
-
-    val deferred = deferred<String, Exception>()
-    deployVerticle(verticleClass.java.getName(), promiseResult(deferred))
-    return deferred.promise
+fun  <T : AbstractVerticle> Vertx.promiseDeployVerticle(verticleClass: KClass<T>): Promise<String, Exception> {
+    return withDeferred { deployVerticle(verticleClass.java.getName(), promiseResult(it)) }
 }
 
 /**
- * Deploy a verticle, returning a Promise<deploymentId as String, Exception>
+ * Deploy a verticle
+ *
+ * @param verticleClass reference to Verticle class (Kotlin `KClass`) to deploy
+ * @param options deployment options
+ * @return Promise<deploymentId as String, Exception>
  */
-public fun  <T : AbstractVerticle> Vertx.promiseDeployVerticle(verticleClass: KClass<T>, options: DeploymentOptions): Promise<String, Exception> {
-    VertxInit.ensure()
-
-    val deferred = deferred<String, Exception>()
-    deployVerticle(verticleClass.java.getName(), options, promiseResult(deferred))
-    return deferred.promise
+fun  <T : AbstractVerticle> Vertx.promiseDeployVerticle(verticleClass: KClass<T>, options: DeploymentOptions): Promise<String, Exception> {
+    return withDeferred { deployVerticle(verticleClass.java.getName(), options, promiseResult(it)) }
 }
 
 /**
- * Deploy a verticle, returning a Promise<deploymentId as String, Exception>
+ * Deploy a verticle
+ *
+ * @param verticleClass reference to Verticle class (Java `Class<*>`) to deploy
+ * @return Promise<deploymentId as String, Exception>
  */
-public fun  <T : AbstractVerticle> Vertx.promiseDeployVerticle(verticleClass: Class<T>): Promise<String, Exception> {
-    VertxInit.ensure()
-
-    val deferred = deferred<String, Exception>()
-    deployVerticle(verticleClass.getName(), promiseResult(deferred))
-    return deferred.promise
+fun  <T : AbstractVerticle> Vertx.promiseDeployVerticle(verticleClass: Class<T>): Promise<String, Exception> {
+    return withDeferred { deployVerticle(verticleClass.getName(), promiseResult(it)) }
 }
 
 /**
- * Deploy a verticle, returning a Promise<deploymentId as String, Exception>
+ * Deploy a verticle
+ *
+ * @param verticleClass reference to Verticle class (Java `Class<*>`) to deploy
+ * @param options deployment options
+ * @return Promise<deploymentId as String, Exception>
  */
-public fun  <T : AbstractVerticle> Vertx.promiseDeployVerticle(verticleClass: Class<T>, options: DeploymentOptions): Promise<String, Exception> {
-    VertxInit.ensure()
-
-    val deferred = deferred<String, Exception>()
-    deployVerticle(verticleClass.getName(), options, promiseResult(deferred))
-    return deferred.promise
+fun  <T : AbstractVerticle> Vertx.promiseDeployVerticle(verticleClass: Class<T>, options: DeploymentOptions): Promise<String, Exception> {
+    return withDeferred { deployVerticle(verticleClass.getName(), options, promiseResult(it)) }
 }
 
 /**
- * Deploy a verticle, returning a Promise<deploymentId as String, Exception>
+ * Deploy a verticle
+ *
+ * @param name verticle by identifier
+ * @return Promise<deploymentId as String, Exception>
  */
-public fun Vertx.promiseDeployVerticle(name: String): Promise<String, Exception> {
-    VertxInit.ensure()
-
-    val deferred = deferred<String, Exception>()
-    deployVerticle(name, promiseResult(deferred))
-    return deferred.promise
+fun Vertx.promiseDeployVerticle(name: String): Promise<String, Exception> {
+    return withDeferred { deployVerticle(name, promiseResult(it)) }
 }
 
 /**
- * Deploy a verticle, returning a Promise<deploymentId as String, Exception>
+ * Deploy a verticle
+ *
+ * @param name verticle by identifier
+ * @param options deployment options
+ * @return Promise<deploymentId as String, Exception>
  */
-public fun Vertx.promiseDeployVerticle(name: String, options: DeploymentOptions): Promise<String, Exception> {
-    VertxInit.ensure()
-
-    val deferred = deferred<String, Exception>()
-    deployVerticle(name, options, promiseResult(deferred))
-    return deferred.promise
+fun Vertx.promiseDeployVerticle(name: String, options: DeploymentOptions): Promise<String, Exception> {
+    return withDeferred { deployVerticle(name, options, promiseResult(it)) }
 }
 
 /**
- * Deploy a verticle async without waiting for it to complete
+ * Deploy a verticle async without waiting for it to complete or tracking it in any way
+ *
+ * @param verticleClass reference to Verticle class (Kotlin `KClass`) to deploy
  */
-public fun <T : AbstractVerticle> Vertx.deployVerticle(verticleClass: KClass<T>): Unit {
+fun <T : AbstractVerticle> Vertx.deployVerticle(verticleClass: KClass<T>): Unit {
+    VertxInit.ensure()
     deployVerticle(verticleClass.java.getName())
 }
 
 /**
- * Deploy a verticle async without waiting for it to complete
+ * Deploy a verticle async without waiting for it to complete or tracking it in any way
+ *
+ * @param verticleClass reference to Verticle class (Java `Class<*>`) to deploy
  */
-public fun <T : AbstractVerticle> Vertx.deployVerticle(verticleClass: Class<T>): Unit {
+fun <T : AbstractVerticle> Vertx.deployVerticle(verticleClass: Class<T>): Unit {
+    VertxInit.ensure()
     deployVerticle(verticleClass.getName())
 }
 
-public class NADA {} // we can't deal with Void return type Vert.x likes to define but really means NULL but isn't nullable
-val nada = NADA()
+/**
+ * Undeploy a verticle
+ *
+ * @return Promise<Unit, Exception> an empty promise that is successful when undeploy completes
+ */
+fun Vertx.promiseUndeploy(deploymentId: String): Promise<Unit, Exception> {
+    return withDeferred { this.undeploy(deploymentId, promiseVoidResult(it)) }
+}
 
 /**
- * Undeploy a verticle, returning a Promise<NADA, Exception>
+ * Close vert.x
+ *
+ * @return Promise<Unit, Exception> an empty promise that is successful when closing completes
  */
-public fun Vertx.promiseUndeploy(deploymentId: String): Promise<NADA, Exception> {
-    VertxInit.ensure()
+fun Vertx.promiseClose(): Promise<Unit, Exception> {
+    return withDeferred { this.close(promiseVoidResult(it)) }
+}
 
-    val deferred = deferred<NADA, Exception>()
-    this.undeploy(deploymentId, { completion ->
-        if (completion.succeeded()) {
-            deferred.resolve(nada)
-        } else {
-            if (completion.cause() is Exception) {
-                deferred.reject(completion.cause() as Exception)
-            } else {
-                deferred.reject(WrappedThrowableException(completion.cause()))
+/**
+ * Execute blocking code using vert.x dispatcher returning a `Promise<T, Exception>`.  Since Kovenant and
+ * vert.x dispatching are united, this is the same as doing `task { ... }` in Kovenant except that no
+ * vert.x context will be in thread local storage if you do not use this method.
+ *
+ * @param blockingCode code to execute
+ * @return Promise<T, Exception> where T is the return type of blockingCode
+ */
+fun <T : Any> Vertx.promiseExecuteBlocking(blockingCode: () -> T): Promise<T, Exception> {
+    return withDeferred { deferred ->
+        this.executeBlocking({ response ->
+            try {
+                response.complete(blockingCode())
+            } catch (ex: Throwable) {
+                response.fail(ex)
             }
-        }
-    })
-    return deferred.promise
+        }, promiseResult(deferred))
+    }
 }
 
 /**
- * Close vertx, returning a Promise<Void, Exception>
+ * Execute blocking code using vert.x dispatcher returning a `Promise<T, Exception>`.  Since Kovenant and
+ * vert.x dispatching are united, this is the same as doing `task { ... }` in Kovenant except that no
+ * vert.x context will be in thread local storage if you do not use this method.
+ *
+ * @param blockingCode code to execute
+ * @return Promise<T, Exception> where T is the return type of blockingCode
  */
-public fun Vertx.promiseClose(): Promise<NADA, Exception> {
-    val deferred = deferred<NADA, Exception>()
-    this.close({ completion ->
-        if (completion.succeeded()) {
-            deferred.resolve(nada)
-        } else {
-            if (completion.cause() is Exception) {
-                deferred.reject(completion.cause() as Exception)
-            } else {
-                deferred.reject(WrappedThrowableException(completion.cause()))
+fun <T : Any> Vertx.executeBlocking(blockingCode: () -> T): Promise<T, Exception> {
+    return withDeferred { deferred ->
+        this.executeBlocking({ response ->
+            try {
+                response.complete(blockingCode())
+            } catch (ex: Throwable) {
+                response.fail(ex)
             }
-        }
-    })
-    return deferred.promise
+        }, promiseResult(deferred))
+    }
 }
 
 /**
- * Execute blocking code using vertx dispatcher returning a Promise<T, Exception>.  Since Kovenant and
- * vertx dispatching are united, this is the same as doing async { ... } in Kovenant, no need to call on a
- * vertx instance.
+ * Sends the message and returns a promise of a reply to that message.
+ *
+ * @param address Message-bus address to send to
+ * @param message The message to send on the message-bus
+ * @return Promise<T, Exception>
  */
-public fun <T: Any> Vertx.promiseExecuteBlocking(blockingCode: () -> T): Promise<T, Exception> {
-    VertxInit.ensure()
-
-    val deferred = deferred<T, Exception>()
-    this.executeBlocking({ response ->
-        try {
-            response.complete(blockingCode())
-        } catch (ex: Throwable) {
-            response.fail(ex)
-        }
-    }, promiseResult(deferred))
-    return deferred.promise
+fun <T : Any> Vertx.promiseReply(address: String, message: Any): Promise<T, Exception> {
+    return withDeferred { deferred ->
+        eventBus().send(address, message, Handler { ar: AsyncResult<Message<T>> ->
+            if (ar.succeeded()) {
+                deferred.resolve(ar.result().body())
+            } else {
+                deferred.reject(Exception(ar.cause()))
+            }
+        })
+    }
 }
 
 /**
- * Execute blocking code using vertx dispatcher returning a Promise<T, Exception>.  Since Kovenant and
- * vertx dispatching are united, this is the same as doing async { ... } in Kovenant, no need to call on a
- * vertx instance.
+ * Setup Kovenant do dispatch via Vert.x, and ensure the Vert.x Jackson object mapper is setup for Kotlin and JDK 8 types.
+ *
+ * All promise verison of startup methods ensure this object is initialized before continuing.  If you are starting Vert.x
+ * in some other way, you should call [`VertxInit.ensure()`][VertxInit.ensure] before doing anything with Vert.x + Kovenant
+ * together.  Calling the method more than once does nothing, and is not harmful (it only causes the class to load if not loaded).
  */
-public fun <T: Any> Vertx.executeBlocking(blockingCode: () -> T): Promise<T, Exception> {
-    VertxInit.ensure()
-
-    val deferred = deferred<T, Exception>()
-    this.executeBlocking({ response ->
-        try {
-            response.complete(blockingCode())
-        } catch (ex: Throwable) {
-            response.fail(ex)
-        }
-    }, promiseResult(deferred))
-    return deferred.promise
-}
-
-
-/**
- * Setup Kovenant do dispatch via Vert-x, and ensure the Vert.x Jackson object mapper is setup for Kotlin and JDK 8 types
- */
-
-public object VertxInit {
+object VertxInit {
     val fallbackContext = Kovenant.createContext {
         workerContext.dispatcher {
             name = "worker-fallback"
@@ -292,8 +313,11 @@ public object VertxInit {
 
     }
 
+    /**
+     * Be sure that Kovenant and Vert.x are initialized in a way that they cooperate and work together.
+     */
     @Suppress("NOTHING_TO_INLINE")
-    public inline fun ensure() {
+    inline fun ensure() {
         // TODO: here to be sure we have intiailized anything related before using,
         //       although this function may remain empty it causes initializers on the
         //       object to run.
@@ -301,16 +325,29 @@ public object VertxInit {
 }
 
 /**
- * Helper to convert an expectation of AsyncResult<T> into a promise represented by Deferred<T, Throwable>
+ * Helper to convert an expectation of `AsyncResult<T>` into a promise represented by `Deferred<T, Exception>`
  *
- *     i.e.
+ *     for example:
+ *
+ *       ```
  *       public fun someAsynActionAsPromise(): Promise<SomeType, Exception> {
  *           val deferred = deferred<SomeType, Exception>()
  *           vertx.someAsyncAction( promiseResult(deferred) )
  *           return deferred.promise
  *       }
+ *       ```
+ *
+ *     or a shorter version:
+ *
+ *       ```
+ *       public fun someAsynActionAsPromise(): Promise<SomeType, Exception> {
+ *           return withDeferred { vertx.someAsyncAction( promiseResult(it) ) }
+ *       }
+ *       ```
  */
-public fun <T: Any> promiseResult(deferred: Deferred<T, Exception>): (AsyncResult<T>) -> Unit {
+fun <T : Any?> promiseResult(deferred: Deferred<T, Exception>): (AsyncResult<T>) -> Unit {
+    VertxInit.ensure()
+
     return { completion ->
         if (completion.succeeded()) {
             deferred.resolve(completion.result())
@@ -325,19 +362,43 @@ public fun <T: Any> promiseResult(deferred: Deferred<T, Exception>): (AsyncResul
 }
 
 /**
- * Sends the message and returns a promise of a reply to that message.
+ * Same as [promiseResult] but to be used when the `AsyncResult` type is `Void`
  */
-public fun <T> Vertx.replyPromise(address: String, message: Any): Promise<T, Exception> {
+fun promiseVoidResult(deferred: Deferred<Unit, Exception>): (AsyncResult<Void>) -> Unit {
+    VertxInit.ensure()
+
+    return { completion ->
+        if (completion.succeeded()) {
+            deferred.resolve(Unit)
+        } else {
+            if (completion.cause() is Exception) {
+                deferred.reject(completion.cause() as Exception)
+            } else {
+                deferred.reject(WrappedThrowableException(completion.cause()))
+            }
+        }
+    }
+}
+
+/**
+ * Helper function that creates a deferred for a block of code and returns the promise associated with the deferred
+ *
+ *     for example:
+ *
+ *       ```
+ *       public fun someAsynActionAsPromise(): Promise<SomeType, Exception> {
+ *           return withDeferred { vertx.someAsyncAction( promiseResult(it) ) }
+ *       }
+ *       ```
+ *
+ * @param codeBlock the block of code that is executed and will resolve or reject the `deferred`
+ * @return Promise<T, Exception> where `T` is the return type of the codeBlock
+ */
+fun <T : Any?> withDeferred(codeBlock: (deferred: Deferred<T, Exception>) -> Unit): Promise<T, Exception> {
     VertxInit.ensure()
 
     val deferred = deferred<T, Exception>()
-    eventBus().send(address, message, Handler { ar: AsyncResult<Message<T>> ->
-        if (ar.succeeded()) {
-            deferred.resolve(ar.result().body())
-        } else {
-            deferred.reject(Exception(ar.cause()))
-        }
-    })
-
+    codeBlock(deferred)
     return deferred.promise
 }
+