+++ vertx3-jdk8/src/main/kotlin/uy/klutter/vertx/json/VertxJson.kt
@@ -1,3 +1,5 @@
+@file:Suppress("NOTHING_TO_INLINE")
+
 package uy.klutter.vertx.json
 
 import io.vertx.core.json.Json
@@ -6,68 +8,114 @@ import io.vertx.core.json.JsonObject
 import uy.klutter.core.jdk8.toIsoString
 import java.time.temporal.Temporal
 
-
-@Suppress("NOTHING_TO_INLINE")
-public inline fun jsonObjectFromString(json: String): JsonObject {
+/**
+ * converts a JSON formatted string containing an object into a tree of JsonObject
+ */
+inline fun jsonObjectFromString(json: String): JsonObject {
     return JsonObject(json)
 }
 
-@Suppress("NOTHING_TO_INLINE")
-public inline fun jsonArrayFromString(json: String): JsonArray {
+/**
+ * converts a JSON formatted string containing an array into an JsonArray
+ */
+inline fun jsonArrayFromString(json: String): JsonArray {
     return JsonArray(json)
 }
 
-@Suppress("NOTHING_TO_INLINE")
-public inline fun <V> jsonObjectFromMap(map: Map<String, V>): JsonObject {
+/**
+ * Converts a Map<String, T> into a JsonObject
+ */
+inline fun <V> jsonObjectFromMap(map: Map<String, V>): JsonObject {
     return JsonObject(map)
 }
 
-@Suppress("NOTHING_TO_INLINE")
-public inline fun <T> jsonArrayFromList(list: List<T>): JsonArray {
+/**
+ * Converts a List<T> into a JsonArray
+ */
+inline fun <T> jsonArrayFromList(list: List<T>): JsonArray {
     return JsonArray(list)
 }
 
-@Suppress("NOTHING_TO_INLINE", "UNCHECKED_CAST", "PLATFORM_CLASS_MAPPED_TO_KOTLIN")
-public inline fun jsonObjectFromPojo(something: Any): JsonObject {
+/**
+ * Converts a POJO into a JsonObject
+ */
+@Suppress("UNCHECKED_CAST", "PLATFORM_CLASS_MAPPED_TO_KOTLIN")
+inline fun jsonObjectFromPojo(something: Any): JsonObject {
     return jsonObjectFromMap<Any?>(Json.mapper.convertValue(something, java.util.Map::class.java) as Map<String, Any?>)
 }
 
-public inline fun jsonObject(init: JsonObject.() -> Unit): JsonObject {
+/**
+ * A builder function to start creating a JsonObject
+ *
+ * ```
+ * jsonObject {
+ *     put("parm", "value")
+ *     put("field2", 123)
+ * }
+ * ```
+ */
+inline fun jsonObject(init: JsonObject.() -> Unit): JsonObject {
     val jsonObject = JsonObject()
     jsonObject.init()
     return jsonObject
 }
 
-public inline fun jsonArray(init: JsonArray.() -> Unit): JsonArray {
+/**
+ * A builder function to start creating a JsonArray
+ *
+ * ```
+ * jsonArray {
+ *     add("value1")
+ *     add("value2")
+ * }
+ * ```
+ */
+inline fun jsonArray(init: JsonArray.() -> Unit): JsonArray {
     val jsonArray = JsonArray()
     jsonArray.init()
     return jsonArray
 }
 
-public inline fun JsonObject.putObject(name: String, init: JsonObject.() -> Unit): JsonObject {
+/**
+ * A builder function to nest JSON objects within other objects
+ */
+inline fun JsonObject.putObject(name: String, init: JsonObject.() -> Unit): JsonObject {
     put(name, jsonObject(init))
     return this
 }
 
-public inline fun JsonObject.putArray(name: String, init: JsonArray.() -> Unit): JsonObject {
+/**
+ * A builder function to nest JSON arrays within other objects
+ */
+inline fun JsonObject.putArray(name: String, init: JsonArray.() -> Unit): JsonObject {
     put(name, jsonArray(init))
     return this
 }
 
-public inline fun JsonArray.addObject(init: JsonObject.() -> Unit): JsonArray {
+/**
+ * A builder function to nest JSON objects within an array
+ */
+inline fun JsonArray.addObject(init: JsonObject.() -> Unit): JsonArray {
     add(jsonObject(init))
     return this
 }
 
-public inline fun JsonArray.addArray(init: JsonArray.() -> Unit): JsonArray {
+/**
+ * A builder function to nest JSON arrays within an array
+ */
+inline fun JsonArray.addArray(init: JsonArray.() -> Unit): JsonArray {
     add(jsonArray(init))
     return this
 }
 
-@Suppress("NOTHING_TO_INLINE")
-public inline fun JsonObject.putDateIsoString(name: String, value: Temporal): JsonObject = put(name, value.toIsoString())
+/**
+ * Put a date into a JsonObject as an ISO formated string
+ */
+inline fun JsonObject.putDateIsoString(name: String, value: Temporal): JsonObject = put(name, value.toIsoString())
 
-@Suppress("NOTHING_TO_INLINE")
-public inline fun JsonArray.addDateIsoString(value: Temporal): JsonArray = add(value.toIsoString())
+/**
+ * Put a date into a JsonArray as an ISO formated string
+ */
+inline fun JsonArray.addDateIsoString(value: Temporal): JsonArray = add(value.toIsoString())